move backend creation and skip tests for unsupported backend
