+++ dlt/destinations/dataset.py
@@ -242,10 +242,13 @@ class ReadableDBAPIDataset(SupportsReadableDataset):
 
     def ibis(self) -> IbisBackend:
         """return a connected ibis backend"""
+        from dlt.common.libs.ibis import create_ibis_backend
+
+        self._ensure_client_and_schema()
         return create_ibis_backend(
             self._destination,
             self._dataset_name,
-            schema=self.schema,
+            self._destination_client(self.schema),
         )
 
     @property
@@ -325,50 +328,6 @@ def dataset(
     if dataset_type == "dbapi":
         return ReadableDBAPIDataset(destination, dataset_name, schema)
     raise NotImplementedError(f"Dataset of type {dataset_type} not implemented")
-
-
-from dlt.common.destination.reference import JobClientBase
-
-
-def create_ibis_backend(
-    destination: TDestinationReferenceArg, dataset_name: str, schema: Schema
-) -> IbisBackend:
-    from dlt.common.libs.ibis import ibis
-    import duckdb
-    from dlt.destinations.impl.duckdb.factory import DuckDbCredentials
-
-    destination = Destination.from_reference(destination)
-    client = _get_client_for_destination(destination, schema, dataset_name)
-
-    if destination.destination_type in ["dlt.destinations.postgres", "dlt.destinations.duckdb"]:
-        credentials = client.config.credentials.to_native_representation()
-        ibis = ibis.connect(credentials)
-    elif destination.destination_type == "dlt.destinations.filesystem":
-        from dlt.destinations.impl.filesystem.sql_client import (
-            FilesystemClient,
-            FilesystemSqlClient,
-        )
-
-        # we create an in memory duckdb and create all tables on there
-        duck = duckdb.connect(":memory:")
-        fs_client = cast(FilesystemClient, client)
-        creds = DuckDbCredentials(duck)
-        sql_client = FilesystemSqlClient(
-            fs_client, dataset_name=fs_client.dataset_name, credentials=creds
-        )
-
-        # NOTE: we should probably have the option for the user to only select a subset of tables here
-        with sql_client as _:
-            sql_client.create_views_for_all_tables()
-        ibis = ibis.duckdb.from_connection(duck)
-
-    else:
-        raise NotImplementedError()
-
-    # NOTE: there seems to be no standardized way to set the current dataset / schema in ibis
-    ibis.raw_sql(f"SET search_path TO {dataset_name};")
-
-    return ibis
 
 
 # helpers
